+++ src/main/java/com/idea5/four_cut_photos_map/domain/review/service/ReviewReadService.java
@@ -2,12 +2,8 @@ package com.idea5.four_cut_photos_map.domain.review.service;
 
 import com.idea5.four_cut_photos_map.domain.member.entity.Member;
 import com.idea5.four_cut_photos_map.domain.memberTitle.service.MemberTitleService;
-import com.idea5.four_cut_photos_map.domain.review.dto.request.RequestReviewDto;
 import com.idea5.four_cut_photos_map.domain.review.dto.response.*;
 import com.idea5.four_cut_photos_map.domain.review.entity.Review;
-import com.idea5.four_cut_photos_map.domain.review.entity.score.ItemScore;
-import com.idea5.four_cut_photos_map.domain.review.entity.score.PurityScore;
-import com.idea5.four_cut_photos_map.domain.review.entity.score.RetouchScore;
 import com.idea5.four_cut_photos_map.domain.review.mapper.ReviewMapper;
 import com.idea5.four_cut_photos_map.domain.review.repository.ReviewRepository;
 import com.idea5.four_cut_photos_map.domain.shop.entity.Shop;
@@ -24,22 +20,13 @@ import java.util.stream.Collectors;
 
 @Slf4j
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class ReviewService {
+public class ReviewReadService {
     private final MemberTitleService memberTitleService;
     private final ReviewRepository reviewRepository;
     private final ShopService shopService;
 
-    private boolean actorCanModify(Member member, Review review) {
-        return member.getId() == review.getWriter().getId();
-    }
-
-    private boolean actorCanDelete(Member member, Review review) {
-        return actorCanModify(member, review);
-    }
-
-    @Transactional(readOnly = true)
     public ResponseReviewDto getReviewById(Long reviewId) {
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.REVIEW_NOT_FOUND));
@@ -47,7 +34,6 @@ public class ReviewService {
         return ReviewMapper.toResponseReviewDto(review);
     }
 
-    @Transactional(readOnly = true)
     public List<ResponseShopReviewDto> getAllShopReviews(Long shopId) {
         Shop shop = shopService.findById(shopId);
 
@@ -58,7 +44,6 @@ public class ReviewService {
     }
 
     // 지점 전체 리뷰 조회
-    @Transactional(readOnly = true)
     public List<ShopReviewResp> getAllShopReview(Long shopId) {
         Shop shop = shopService.findById(shopId);
 
@@ -71,7 +56,6 @@ public class ReviewService {
                 .collect(Collectors.toList());
     }
 
-    @Transactional(readOnly = true)
     public List<ResponseMemberReviewDto> getAllMemberReviews(Long memberId) {
         List<Review> reviews = reviewRepository.findAllByWriterIdOrderByCreateDateDesc(memberId);
 
@@ -80,8 +64,6 @@ public class ReviewService {
                 .collect(Collectors.toList());
     }
 
-
-    @Transactional(readOnly = true)
     public List<ResponseShopReviewDto> getTop3ShopReviews(Long shopId) {
         List<Review> reviews = reviewRepository.findTop3ByShopIdOrderByCreateDateDesc(shopId);
 
@@ -91,7 +73,6 @@ public class ReviewService {
     }
 
     // 리뷰 최신순 3개 조회
-    @Transactional(readOnly = true)
     public List<ShopReviewResp> getTop3ShopReview(Long shopId) {
         List<Review> reviews = reviewRepository.findTop3ByShopIdOrderByCreateDateDesc(shopId);
 
@@ -103,52 +84,11 @@ public class ReviewService {
                 .collect(Collectors.toList());
     }
 
-    public ResponseReviewDto write(Member member, Long shopId, RequestReviewDto reviewDto) {
-        Shop shop = shopService.findById(shopId);
-
-        Review savedReview = reviewRepository.save(ReviewMapper.toEntity(member, shop, reviewDto));
-
-        return ReviewMapper.toResponseReviewDto(savedReview);
-    }
-
-    public ResponseReviewDto modify(Member member, Long reviewId, RequestReviewDto reviewDto) {
-        Review review = reviewRepository.findById(reviewId).orElseThrow(() -> new BusinessException(ErrorCode.REVIEW_NOT_FOUND));
-
-        // 수정 권한 확인
-        if(!actorCanModify(member, review)) {
-            throw new BusinessException(ErrorCode.WRITER_DOES_NOT_MATCH);
-        }
-
-        ReviewMapper.update(review, reviewDto);
-
-        return ReviewMapper.toResponseReviewDto(review);
-    }
-
-    public Long delete(Member member, Long reviewId) {
-        Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.REVIEW_NOT_FOUND));
-        
-        if(!actorCanDelete(member, review)) {
-            throw new BusinessException(ErrorCode.WRITER_DOES_NOT_MATCH);
-        }
-
-        Long shopId = review.getShop().getId();
-
-        reviewRepository.delete(review);
-
-        return shopId;
-    }
-
     // 회원의 리뷰수 조회
     public Long getReviewCntByWriter(Member member) {
         return reviewRepository.countByWriter(member);
     }
 
-    public void deleteByWriterId(Long memberId) {
-        reviewRepository.deleteByWriterId(memberId);
-    }
-
-    @Transactional(readOnly = true)
     public ShopReviewInfoDto getShopReviewInfo(Long shopId) {
         Shop shop = shopService.findById(shopId);
