+++ src/test/java/com/idea5/four_cut_photos_map/domain/review/unit/service/ReviewReadServiceTest.java
@@ -1,9 +1,7 @@
 package com.idea5.four_cut_photos_map.domain.review.unit.service;
 
 import com.idea5.four_cut_photos_map.domain.brand.entity.Brand;
-import com.idea5.four_cut_photos_map.domain.brand.entity.MajorBrand;
 import com.idea5.four_cut_photos_map.domain.member.entity.Member;
-import com.idea5.four_cut_photos_map.domain.review.dto.request.RequestReviewDto;
 import com.idea5.four_cut_photos_map.domain.review.dto.response.ResponseMemberReviewDto;
 import com.idea5.four_cut_photos_map.domain.review.dto.response.ResponseReviewDto;
 import com.idea5.four_cut_photos_map.domain.review.dto.response.ResponseShopReviewDto;
@@ -14,9 +12,8 @@ import com.idea5.four_cut_photos_map.domain.review.entity.score.PurityScore;
 import com.idea5.four_cut_photos_map.domain.review.entity.score.RetouchScore;
 import com.idea5.four_cut_photos_map.domain.review.mapper.ReviewMapper;
 import com.idea5.four_cut_photos_map.domain.review.repository.ReviewRepository;
-import com.idea5.four_cut_photos_map.domain.review.service.ReviewService;
+import com.idea5.four_cut_photos_map.domain.review.service.ReviewReadService;
 import com.idea5.four_cut_photos_map.domain.shop.entity.Shop;
-import com.idea5.four_cut_photos_map.domain.shop.repository.ShopRepository;
 import com.idea5.four_cut_photos_map.domain.shop.service.ShopService;
 import com.idea5.four_cut_photos_map.global.error.ErrorCode;
 import com.idea5.four_cut_photos_map.global.error.exception.BusinessException;
@@ -24,9 +21,7 @@ import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.mockito.stubbing.OngoingStubbing;
 
 import java.time.LocalDateTime;
 import java.util.*;
@@ -36,9 +31,9 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-public class ReviewServiceTest {
+public class ReviewReadServiceTest {
     @InjectMocks
-    private ReviewService reviewService;
+    private ReviewReadService reviewReadService;
     @Mock
     private ReviewRepository reviewRepository;
     @Mock
@@ -72,7 +67,7 @@ public class ReviewServiceTest {
                 // when
                 when(reviewRepository.findById(retrieveReviewId)).thenReturn(Optional.of(review));
 
-                ResponseReviewDto responseReviewDto = reviewService.getReviewById(retrieveReviewId);
+                ResponseReviewDto responseReviewDto = reviewReadService.getReviewById(retrieveReviewId);
 
                 // then
                 Assertions.assertEquals(responseReviewDto.getReviewInfo().getId(), review.getId());
@@ -102,178 +97,7 @@ public class ReviewServiceTest {
                 when(reviewRepository.findById(retrieveReviewId)).thenThrow(exception);
 
                 // then
-                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewService.getReviewById(retrieveReviewId));
-                Assertions.assertEquals(resultException.getMessage(), exception.getMessage());
-            }
-        }
-    }
-
-    @Nested
-    @DisplayName("특정 리뷰 수정")
-    class ModifyReview {
-        private Member writer;
-        private Brand brand;
-        private Shop shop;
-        private Review review;
-
-        @BeforeEach
-        void setUp() {
-            writer = Member.builder().id(1L).kakaoId(1000L).nickname("user1").build();
-            brand = Brand.builder().id(1L).brandName("인생네컷").filePath("https://d18tllc1sxg8cp.cloudfront.net/brand_image/brand_1.jpg").build();
-            shop = Shop.builder().id(1L).brand(brand).placeName("인생네컷망리단길점").address("서울 마포구 포은로 109-1").favoriteCnt(0).reviewCnt(0).starRatingAvg(0.0).build();
-            review = Review.builder().id(1L).createDate(LocalDateTime.now()).modifyDate(LocalDateTime.now()).writer(writer).shop(shop).starRating(5).content("리뷰 내용").purity(PurityScore.BAD).retouch(RetouchScore.BAD).item(ItemScore.BAD).build();
-        }
-
-        @Nested
-        @DisplayName("성공")
-        class SuccessCase {
-            @Test
-            @DisplayName("해당 id 가진 리뷰 수정")
-            void modifyReviewSuccess1() {
-                // given
-                Long modifyReviewId = 1L;
-                Member user = Member.builder().id(1L).build();
-                RequestReviewDto modifyReviewDto = RequestReviewDto.builder().starRating(3).content("수정 후 리뷰 내용").purity("GOOD").retouch("GOOD").item("GOOD").build();
-
-                // when
-                when(reviewRepository.findById(modifyReviewId)).thenReturn(Optional.of(review));
-
-                ResponseReviewDto responseReviewDto = reviewService.modify(user, modifyReviewId, modifyReviewDto);
-
-                // then
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getStarRating(), modifyReviewDto.getStarRating());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getContent(), modifyReviewDto.getContent());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getPurity(), PurityScore.valueOf(modifyReviewDto.getPurity()));
-                Assertions.assertEquals(String.valueOf(responseReviewDto.getReviewInfo().getRetouch()), modifyReviewDto.getRetouch());
-                Assertions.assertEquals(String.valueOf(responseReviewDto.getReviewInfo().getItem()), modifyReviewDto.getItem());
-            }
-
-            @Test
-            @DisplayName("purity, retouch, item를 null로 전송")
-            void modifyReviewSuccess2() {
-                // given
-                Long modifyReviewId = 1L;
-                Member user = Member.builder().id(1L).build();
-                RequestReviewDto modifyReviewDto = RequestReviewDto.builder().starRating(3).content("수정 후 리뷰 내용").build();
-
-                // when
-                when(reviewRepository.findById(modifyReviewId)).thenReturn(Optional.of(review));
-
-                ResponseReviewDto responseReviewDto = reviewService.modify(user, modifyReviewId, modifyReviewDto);
-
-                // then
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getId(), modifyReviewId);
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getStarRating(), modifyReviewDto.getStarRating());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getContent(), modifyReviewDto.getContent());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getPurity(), PurityScore.UNSELECTED);
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getRetouch(), RetouchScore.UNSELECTED);
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getItem(), ItemScore.UNSELECTED);
-
-                Assertions.assertEquals(responseReviewDto.getMemberInfo().getId(), writer.getId());
-                Assertions.assertEquals(responseReviewDto.getMemberInfo().getNickname(), writer.getNickname());
-
-                Assertions.assertEquals(responseReviewDto.getShopInfo().getId(), shop.getId());
-                Assertions.assertEquals(responseReviewDto.getShopInfo().getPlaceName(), shop.getPlaceName());
-            }
-        }
-
-        @Nested
-        @DisplayName("실패")
-        class FailCase {
-            @Test
-            @DisplayName("해당 id 가진 리뷰 존재하지 않음")
-            void modifyReviewFail1() {
-                // given
-                Long modifyReviewId = 2L;
-                Member user = Member.builder().id(1L).build();
-                RequestReviewDto modifyReviewDto = RequestReviewDto.builder().starRating(3).content("수정 후 리뷰 내용").purity("GOOD").retouch("GOOD").item("GOOD").build();
-                BusinessException exception = new BusinessException(ErrorCode.REVIEW_NOT_FOUND);
-
-                // when
-                when(reviewRepository.findById(modifyReviewId)).thenThrow(exception);
-
-                // then
-                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewService.modify(user, modifyReviewId, modifyReviewDto));
-                Assertions.assertEquals(resultException.getErrorCode(), exception.getErrorCode());
-                Assertions.assertEquals(resultException.getMessage(), exception.getMessage());
-            }
-
-            @Test
-            @DisplayName("사용자와 리뷰 작성자 불일치")
-            void modifyReviewFail2() {
-                // given
-                Long modifyReviewId = 1L;
-                Member user = Member.builder().id(2L).build();
-                RequestReviewDto modifyReviewDto = RequestReviewDto.builder().starRating(3).content("수정 후 리뷰 내용").purity("GOOD").retouch("GOOD").item("GOOD").build();
-                BusinessException exception = new BusinessException(ErrorCode.WRITER_DOES_NOT_MATCH);
-
-                // when
-                when(reviewRepository.findById(modifyReviewId)).thenReturn(Optional.of(review));
-
-                // then
-                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewService.modify(user, modifyReviewId, modifyReviewDto));
-                Assertions.assertEquals(resultException.getErrorCode(), exception.getErrorCode());
-                Assertions.assertEquals(resultException.getMessage(), exception.getMessage());
-            }
-        }
-    }
-
-    @Nested
-    @DisplayName("특정 리뷰 삭제")
-    class DeleteReview {
-        private Member writer;
-        private Brand brand;
-        private Shop shop;
-        private Review review;
-
-        @BeforeEach
-        void setUp() {
-            writer = Member.builder().id(1L).kakaoId(1000L).nickname("user1").build();
-            brand = Brand.builder().id(1L).brandName("인생네컷").filePath("https://d18tllc1sxg8cp.cloudfront.net/brand_image/brand_1.jpg").build();
-            shop = Shop.builder().id(1L).brand(brand).placeName("인생네컷망리단길점").address("서울 마포구 포은로 109-1").favoriteCnt(0).reviewCnt(0).starRatingAvg(0.0).build();
-            review = Review.builder().id(1L).createDate(LocalDateTime.now()).modifyDate(LocalDateTime.now()).writer(writer).shop(shop).starRating(5).content("리뷰 내용").purity(PurityScore.UNSELECTED).retouch(RetouchScore.UNSELECTED).item(ItemScore.UNSELECTED).build();
-        }
-
-        @Nested
-        @DisplayName("성공")
-        class SuccessCase {
-
-        }
-
-        @Nested
-        @DisplayName("실패")
-        class FailCase {
-            @Test
-            @DisplayName("해당 id 가진 리뷰 존재하지 않음")
-            void modifyReviewFail1() {
-                // given
-                Long deleteReviewId = 2L;
-                Member user = Member.builder().id(1L).build();
-                BusinessException exception = new BusinessException(ErrorCode.REVIEW_NOT_FOUND);
-
-                // when
-                when(reviewRepository.findById(deleteReviewId)).thenThrow(exception);
-
-                // then
-                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewService.delete(user, deleteReviewId));
-                Assertions.assertEquals(resultException.getErrorCode(), exception.getErrorCode());
-                Assertions.assertEquals(resultException.getMessage(), exception.getMessage());
-            }
-
-            @Test
-            @DisplayName("사용자와 리뷰 작성자 불일치")
-            void modifyReviewFail2() {
-                // given
-                Long modifyReviewId = 1L;
-                Member user = Member.builder().id(2L).build();
-                BusinessException exception = new BusinessException(ErrorCode.WRITER_DOES_NOT_MATCH);
-
-                // when
-                when(reviewRepository.findById(modifyReviewId)).thenReturn(Optional.of(review));
-
-                // then
-                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewService.delete(user, modifyReviewId));
-                Assertions.assertEquals(resultException.getErrorCode(), exception.getErrorCode());
+                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewReadService.getReviewById(retrieveReviewId));
                 Assertions.assertEquals(resultException.getMessage(), exception.getMessage());
             }
         }
@@ -317,7 +141,7 @@ public class ReviewServiceTest {
                 // when
                 when(reviewRepository.findAllByWriterIdOrderByCreateDateDesc(memberId)).thenReturn(reviews);
 
-                List<ResponseMemberReviewDto> memberReviews = reviewService.getAllMemberReviews(memberId);
+                List<ResponseMemberReviewDto> memberReviews = reviewReadService.getAllMemberReviews(memberId);
 
                 // then
                 Assertions.assertEquals(memberReviews.size(), 2);
@@ -344,7 +168,7 @@ public class ReviewServiceTest {
                 // when
                 when(reviewRepository.findAllByWriterIdOrderByCreateDateDesc(memberId)).thenReturn(reviews);
 
-                List<ResponseMemberReviewDto> memberReviews = reviewService.getAllMemberReviews(memberId);
+                List<ResponseMemberReviewDto> memberReviews = reviewReadService.getAllMemberReviews(memberId);
 
                 // then
                 Assertions.assertEquals(memberReviews.size(), 1);
@@ -366,7 +190,7 @@ public class ReviewServiceTest {
                 // when
                 when(reviewRepository.findAllByWriterIdOrderByCreateDateDesc(memberId)).thenReturn(reviews);
 
-                List<ResponseMemberReviewDto> memberReviews = reviewService.getAllMemberReviews(memberId);
+                List<ResponseMemberReviewDto> memberReviews = reviewReadService.getAllMemberReviews(memberId);
 
                 // then
                 Assertions.assertEquals(memberReviews.size(), 0);
@@ -424,7 +248,7 @@ public class ReviewServiceTest {
                 when(shopService.findById(shopId)).thenReturn(shop1);
                 when(reviewRepository.findAllByShopIdOrderByCreateDateDesc(shopId)).thenReturn(reviews);
 
-                List<ResponseShopReviewDto> shopReviews = reviewService.getAllShopReviews(shopId);
+                List<ResponseShopReviewDto> shopReviews = reviewReadService.getAllShopReviews(shopId);
 
                 // then
                 Assertions.assertEquals(shopReviews.size(), reviews.size());
@@ -452,7 +276,7 @@ public class ReviewServiceTest {
                 when(shopService.findById(shopId)).thenReturn(shop2);
                 when(reviewRepository.findAllByShopIdOrderByCreateDateDesc(shopId)).thenReturn(reviews);
 
-                List<ResponseShopReviewDto> shopReviews = reviewService.getAllShopReviews(shopId);
+                List<ResponseShopReviewDto> shopReviews = reviewReadService.getAllShopReviews(shopId);
 
                 // then
                 Assertions.assertEquals(shopReviews.size(), reviews.size());
@@ -474,7 +298,7 @@ public class ReviewServiceTest {
                 when(shopService.findById(shopId)).thenReturn(shop3);
                 when(reviewRepository.findAllByShopIdOrderByCreateDateDesc(shopId)).thenReturn(reviews);
 
-                List<ResponseShopReviewDto> shopReviews = reviewService.getAllShopReviews(shopId);
+                List<ResponseShopReviewDto> shopReviews = reviewReadService.getAllShopReviews(shopId);
 
                 // then
                 Assertions.assertEquals(shopReviews.size(), reviews.size());
@@ -495,116 +319,7 @@ public class ReviewServiceTest {
                 when(shopService.findById(shopId)).thenThrow(exception);
 
                 // then
-                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewService.getAllShopReviews(shopId));
-                Assertions.assertEquals(resultException.getErrorCode(), exception.getErrorCode());
-                Assertions.assertEquals(resultException.getMessage(), exception.getMessage());
-            }
-        }
-    }
-
-    @Nested
-    @DisplayName("상점 리뷰 작성")
-    class WriteReviewShop {
-        private Member writer;
-        private Brand brand;
-        private Shop shop;
-
-        @BeforeEach
-        void setUp() {
-            writer = Member.builder().id(1L).kakaoId(1000L).nickname("user1").build();
-            brand = Brand.builder().id(1L).brandName("인생네컷").filePath("https://d18tllc1sxg8cp.cloudfront.net/brand_image/brand_1.jpg").build();
-            shop = Shop.builder().id(1L).brand(brand).placeName("인생네컷망리단길점").address("서울 마포구 포은로 109-1").favoriteCnt(0).reviewCnt(0).starRatingAvg(0.0).build();
-        }
-
-        @Nested
-        @DisplayName("성공")
-        class SuccessCase {
-            @Test
-            @DisplayName("shopId 가진 지점에 review 추가")
-            void writeReviewShopSuccess1() {
-                // given
-                Long shopId = 1L;
-                RequestReviewDto requestReviewDto = RequestReviewDto.builder().starRating(3).content("새로 지점에 추가하는 리뷰 내용").purity("GOOD").retouch("GOOD").item("GOOD").build();
-                Review review = Review.builder()
-                        .id(1L)
-                        .createDate(LocalDateTime.now())
-                        .modifyDate(LocalDateTime.now())
-                        .writer(writer)
-                        .shop(shop)
-                        .starRating(requestReviewDto.getStarRating())
-                        .content(requestReviewDto.getContent())
-                        .purity(PurityScore.valueOf(requestReviewDto.getPurity()))
-                        .retouch(RetouchScore.valueOf(requestReviewDto.getRetouch()))
-                        .item(ItemScore.valueOf(requestReviewDto.getItem()))
-                        .build();
-
-                // when
-                when(shopService.findById(shopId)).thenReturn(shop);
-                when(reviewRepository.save(any(Review.class))).thenReturn(review);
-
-                ResponseReviewDto responseReviewDto = reviewService.write(writer, shopId, requestReviewDto);
-
-                // then
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getId(), review.getId());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getStarRating(), review.getStarRating());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getContent(), review.getContent());
-
-            }
-
-            @Test
-            @DisplayName("shopId 가진 지점에 purity, retouch, item null인 review 추가")
-            void retrieveShopReviewsSuccess2() {
-                // given
-                Long shopId = 1L;
-                RequestReviewDto requestReviewDto = RequestReviewDto.builder().starRating(3).content("새로 지점에 추가하는 리뷰 내용").build();
-                Review review = Review.builder()
-                        .id(1L)
-                        .createDate(LocalDateTime.now())
-                        .modifyDate(LocalDateTime.now())
-                        .writer(writer)
-                        .shop(shop)
-                        .starRating(requestReviewDto.getStarRating())
-                        .content(requestReviewDto.getContent())
-                        .purity(PurityScore.UNSELECTED)
-                        .retouch(RetouchScore.UNSELECTED)
-                        .item(ItemScore.UNSELECTED)
-                        .build();
-
-                // when
-                when(shopService.findById(shopId)).thenReturn(shop);
-                when(reviewRepository.save(any(Review.class))).thenReturn(review);
-
-                ResponseReviewDto responseReviewDto = reviewService.write(writer, shopId, requestReviewDto);
-
-                // then
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getId(), review.getId());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getStarRating(), review.getStarRating());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getContent(), review.getContent());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getPurity(), review.getPurity());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getRetouch(), review.getRetouch());
-                Assertions.assertEquals(responseReviewDto.getReviewInfo().getItem(), review.getItem());
-
-                Assertions.assertEquals(responseReviewDto.getShopInfo().getId(), shop.getId());
-                Assertions.assertEquals(responseReviewDto.getShopInfo().getPlaceName(), shop.getPlaceName());
-            }
-        }
-
-        @Nested
-        @DisplayName("실패")
-        class FailCase {
-            @Test
-            @DisplayName("ShopId 존재하지 않는 지점")
-            void retrieveShopReviewsFail1() {
-                // given
-                Long shopId = 2L;
-                RequestReviewDto requestReviewDto = RequestReviewDto.builder().starRating(3).content("새로 지점에 추가하는 리뷰 내용").purity("GOOD").retouch("GOOD").item("GOOD").build();
-                BusinessException exception = new BusinessException(ErrorCode.SHOP_NOT_FOUND);
-
-                // when
-                when(shopService.findById(shopId)).thenThrow(exception);
-
-                // then
-                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewService.write(writer, shopId, requestReviewDto));
+                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewReadService.getAllShopReviews(shopId));
                 Assertions.assertEquals(resultException.getErrorCode(), exception.getErrorCode());
                 Assertions.assertEquals(resultException.getMessage(), exception.getMessage());
             }
@@ -641,7 +356,7 @@ public class ReviewServiceTest {
                 when(reviewRepository.countByShop(shop)).thenReturn(reviewCount);
                 when(reviewRepository.getAverageStarRating(shopId)).thenReturn(starRatingAvg);
 
-                ShopReviewInfoDto shopReviewInfo = reviewService.getShopReviewInfo(shopId);
+                ShopReviewInfoDto shopReviewInfo = reviewReadService.getShopReviewInfo(shopId);
 
                 // then
                 Assertions.assertEquals(shopReviewInfo.getShopId(), shopId);
@@ -661,7 +376,7 @@ public class ReviewServiceTest {
                 when(shopService.findById(shopId)).thenReturn(shop);
                 when(reviewRepository.countByShop(shop)).thenReturn(reviewCount);
 
-                ShopReviewInfoDto shopReviewInfo = reviewService.getShopReviewInfo(shopId);
+                ShopReviewInfoDto shopReviewInfo = reviewReadService.getShopReviewInfo(shopId);
 
                 // then
                 Assertions.assertEquals(shopReviewInfo.getShopId(), shopId);
@@ -684,7 +399,7 @@ public class ReviewServiceTest {
                 when(shopService.findById(shopId)).thenThrow(exception);
 
                 // then
-                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewService.getShopReviewInfo(shopId));
+                BusinessException resultException = Assertions.assertThrows(exception.getClass(), () -> reviewReadService.getShopReviewInfo(shopId));
                 Assertions.assertEquals(resultException.getErrorCode(), exception.getErrorCode());
                 Assertions.assertEquals(resultException.getMessage(), exception.getMessage());
             }
@@ -727,7 +442,7 @@ public class ReviewServiceTest {
                 // when
                 when(reviewRepository.countByWriter(user)).thenReturn(userReviewCnt);
 
-                Long result = reviewService.getReviewCntByWriter(user);
+                Long result = reviewReadService.getReviewCntByWriter(user);
 
                 // then
                 Assertions.assertEquals(userReviewCnt, result);
@@ -743,7 +458,7 @@ public class ReviewServiceTest {
                 // when
                 when(reviewRepository.countByWriter(user)).thenReturn(userReviewCnt);
 
-                Long resultCnt = reviewService.getReviewCntByWriter(user);
+                Long resultCnt = reviewReadService.getReviewCntByWriter(user);
 
                 // then
                 Assertions.assertEquals(userReviewCnt, resultCnt);
@@ -805,7 +520,7 @@ public class ReviewServiceTest {
                 // when
                 when(reviewRepository.findTop3ByShopIdOrderByCreateDateDesc(shopId)).thenReturn(topShopReviews);
 
-                List<ResponseShopReviewDto> resultShopReviewDtoList = reviewService.getTop3ShopReviews(shopId);
+                List<ResponseShopReviewDto> resultShopReviewDtoList = reviewReadService.getTop3ShopReviews(shopId);
 
                 // then
                 Assertions.assertEquals(resultShopReviewDtoList.size(), 3);
@@ -830,7 +545,7 @@ public class ReviewServiceTest {
                 // when
                 when(reviewRepository.findTop3ByShopIdOrderByCreateDateDesc(shopId)).thenReturn(topShopReviews);
 
-                List<ResponseShopReviewDto> resultShopReviewDtoList = reviewService.getTop3ShopReviews(shopId);
+                List<ResponseShopReviewDto> resultShopReviewDtoList = reviewReadService.getTop3ShopReviews(shopId);
 
                 // then
                 Assertions.assertEquals(resultShopReviewDtoList.size(), reviewCnt);
